Move around parse_* functions, some tests
The parse_* functions are now self-standing functions but they are
also available as static methods of B1500 and B1500Module classes.
